+++ packages/flame/src/hooks/useEventListener.ts
@@ -5,21 +5,26 @@ type CallbackEvent<T> = (event: T) => void;
 export function useEventListener<T extends Event>(
   eventKey: string,
   callback: CallbackEvent<T>,
-  target?: EventTarget,
+  // On SSR, document may not be defined at all
+  target: EventTarget | undefined = typeof document !== 'undefined' ? document : undefined,
 ) {
+  // Store callback in a ref so we don't have to remove/add the event listener when callback changes
+  const callbackRef = React.useRef(callback);
   React.useEffect(() => {
-    const captureEvent: CallbackEvent<T> = event => {
-      callback(event);
-    };
-
-    if (target || document) {
-      (target || document).addEventListener(eventKey, captureEvent as any);
-    }
-
-    return () => {
-      if (target || document) {
-        (target || document).removeEventListener(eventKey, captureEvent as any);
-      }
-    };
+    callbackRef.current = callback;
   }, [callback]);
+
+  // We have a conditional return for the cleanup function in this specific case,
+  // we can safely disable this rule.
+  // eslint-disable-next-line consistent-return
+  React.useEffect(() => {
+    const handler: CallbackEvent<any> = event => callbackRef.current(event);
+
+    if (target) {
+      target.addEventListener(eventKey, handler);
+      return () => {
+        target.removeEventListener(eventKey, handler);
+      };
+    }
+  }, [target, eventKey]);
 }