+++ packages/flame/CHANGELOG.md
@@ -7,6 +7,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Refer to the [CONTRIBUTING guide](https://github.com/lightspeed/flame/blob/master/.github/CONTRIBUTING.md) for more info.
 
+## [Unreleased]
+
+### Fixed
+
+- Refactor and improve `useEventListener` hook ([#4](https://github.com/lightspeed/flame/pull/4))
+
 ## 0.1.0 - 2019-09-11
 
 ### Added